requirements.txt
flask==2.3.2
openai==0.27.8
openpyxl==3.0.9
pandas==1.3.4
python-dotenv==1.0.0
requests==2.31.0



